♻️ Refactoring & 📝 Comments
