--- pages/event/edit.html
@@ -1,111 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-
-    <!-- CSS -->
-    <link rel="stylesheet" href="../../src/assets/css/main.css">
-    <link rel="stylesheet" href="../../src/assets/library/daterangepicker/daterangepicker.css">
-
-    <!-- JS -->
-    <script src="../../dist/front/pages/eventAdd.js" type="module" defer></script>
-    <!-- Date Range Picker -->
-    <script type="text/javascript" src="../../src/assets/library/daterangepicker/jquery.min.js"
-        onload="window.$ = window.jQuery = module.exports;"></script>
-    <script type="text/javascript" src="../../src/assets/library/daterangepicker/moment.min.js"></script>
-    <script type="text/javascript" src="../../src/assets/library/daterangepicker/daterangepicker.js"></script>
-</head>
-
-<body>
-
-    <h1>Add Event</h1>
-
-    <form id="addEventForm">
-        <div class="required">
-            <label for="add_event_date">Date and Time </label>
-            <input type="text" name="add_event_date" id="add_event_date" required>
-        </div>
-
-        <div class="required">
-            <label for="add_event_title">Title</label>
-            <input type="text" name="add_event_title" id="add_event_title" required>
-        </div>
-
-        <div>
-            <label for="add_event_location">Location</label>
-            <input type="text" name="add_event_location" id="add_event_location">
-        </div>
-
-        <div>
-            <label for="add_event_category">Category</label>
-            <input type="text" name="add_event_category" id="add_event_category">
-        </div>
-
-        <div>
-            <label for="add_event_status">Status</label>
-            <input type="text" name="add_event_status" id="add_event_status">
-        </div>
-
-        <div>
-            <label for="add_event_description">Description</label>
-            <textarea name="add_event_description" id="add_event_description" rows="10"></textarea>
-        </div>
-
-        <div>
-            <label for="add_event_transparence">Transparence</label>
-            <input type="text" name="add_event_transparence" id="add_event_transparence">
-        </div>
-
-        <input type="submit" value="Add">
-    </form>
-
-
-    <script>
-        $(function () {
-            $('input[name="add_event_date"]').daterangepicker({
-                timePicker: true,
-                startDate: moment().startOf('hour'),
-                endDate: moment().startOf('hour').add(32, 'hour'),
-                timePicker24Hour: true,
-                locale: {
-                    "format": "DD/MM/YYYY HH:mm",
-                    "separator": " - ",
-                    "applyLabel": "Appliquer",
-                    "cancelLabel": "Annuler",
-                    "fromLabel": "De",
-                    "toLabel": "à",
-                    "customRangeLabel": "Custom",
-                    "weekLabel": "W",
-                    "daysOfWeek": [
-                        "Di",
-                        "Lu",
-                        "Ma",
-                        "Me",
-                        "Je",
-                        "Ve",
-                        "Sa"
-                    ],
-                    "monthNames": [
-                        "Janvier",
-                        "Février",
-                        "Mars",
-                        "Avril",
-                        "Mai",
-                        "Juin",
-                        "Juillet",
-                        "Août",
-                        "Septembre",
-                        "Octobre",
-                        "Novembre",
-                        "Décembre"
-                    ],
-                    "firstDay": 1
-                },
-            });
-        });
-    </script>
-</body>
-
-</html>
